Fix corrupted initial prompt
The initial prompt may get corrupted so it looks something like

any/src $ het-vm ~/projects/gea

instead of

techet@techet-vm ~/projects/geany/src $

This is just the initial display problem - when enter is pressed, the
second line already shows correct prompt.

It appears that VTE isn't fully initialized when starting the shell process
and when the shell start is delayed a bit (performed on idle), the prompt
looks correct.