Fix invalid memory access when quitting
When quitting and the prefs dialog has been created, the StashPrefs are
destroyed before the stash tree and so the stash tree cleanup code
accesses freed memory.  Fix this by removing access to the StashPref in
the tree cleanup code, as fixing it properly is a bit of a mess.

Closes #538.