Rename "maintain history on reload" setting temporarily for 1.25
We want it disabled by default in 1.25 for everyone, but want to be
able to get it back on by default later on when the UI is more
polished.

So we will need to have a way to tell whether the configuration comes
from 1.25 and should be upgraded or was willfully disabled by the user
in a later version.  So, use a temporary setting name that defaults to
disabled for 1.25

See #553.