notebooks: Rework switch-page and focus handling
Previously lots of things where triggered when switching tabs. However, with
multiple notebooks one has to consider moving the focus between the notebooks
as another form of switching pages.

The switch-page event is not usable anymore as it's not triggered when the
focus moves between notebooks. The focus-in-event of the tab contents is more
suitable. However, there are times when notebook pages become visible without
receiving focus (e.g. pressing space on the open documents list). Here the map
event comes to the rescue. GeanyPage has combined both to provide a selected
signal that both notebook.c and document.c can use to do their stuff.

Implementation note: The focus changed signals are only emitted for the acutal
focused child (and not GeanyPage). Because of this, GeanyPage has to track the
focus child some smarts in order to also receive those. It does this by simply
by overriding the set-focus-child default handler (and chaining up of course).