Allow plugins to disable GeanyEditor features
If plugins provide replacements for various editor features like
auto-completion, calltips, compiler squiggles, etc, they will
want to turn off the builtin functionality that clashes with
what the plugin provides.

This API is only a stop-gap until a more robust interface can be
designed and implemented which allows plugins to cooperatively
provide backends for some of these  features. In the meantime
this allows development of advanced plugins to continue.