Improve "Open in New Window" feature
Make an attempt to use the same executable file for
the new window as the current instance of Geany. Also
flush configuration to disk and provide path to it
for the new instance using the `-c` option.

Add platform-specific code for Linux, Win32, OSX and
various BSDs. If platform-specific method fails, try
and use the installed path of Geany. If that fails,
try and find `geany` in `PATH` as it did before.