Enhance detection of sh functions
Bash allows more characters besides [[:alnum:]_] when declaring
function names using the 'function' keyword.  It also does not require
having a pair of parentheses after the name.  Some shells may actually
implement it differently but we don't have to be that strict since the
user explicitly specifies the 'function' keyword anyway.

This update implements the one described above, and also invalidates
function names that are completely made up of digits.