C/C++/C#/Java/Vera/Vala/D: Allow non-ASCII identifiers
While it might only currently make sense for Java, it should be
harmless for other languages and is a lot easier and cleaner to
implement like that.

Also, note that this allows more things than Java actually allows, as
it accepts any non-ASCII byte as an identifier, while Java seems to
require a Unicode alphanumeric one.  Again, it should be harmless and
is a lot easier to implement than fully-fledged Unicode support.

Closes #691.