plugins: pass GModule * to geany_load_module()
This is rquired to enable the plugin to call g_module_make_resident() within
its geany_load_module(). This must be done if the plugin registers new
GTypes within that function.

plugin_module_make_resident() doesn't work because it's not an allowed API
function to call from there. Simply allowing it doesn't suffice because
internally the built-in plugin loader isn't fully set up yet:
plugin->proxy_data isn't set to the GModule at this point (by design), and
plugin_module_make_resident() requires that to be set.

This changes a public API, but it is unreleased so far, so not a problem (by
definition).