Update filetypes.c
Added "bool", "true", and "false" to "primary=" in filetypes.c.
This addition was suggested by Colomban Wendling and Devyn Collier Johnson in the mailing list. As Colomban Wendling said -

"This said, we could probably indeed have 'bool', 'false' and 'true' in filetypes.c's 'primary=' list (especially as we do have 'FALSE' and 'TRUE' which aren't even defined in standard C).  Technically those aren't keywords [1], but I guess it'd be convenient to have them -- yet one has to remember those are library extensions only available through stdbool.h.

[1] see i.e. section 6.4.1§1 in the ISO/IEC 9899:201x n1570 draft"

In addition, "bool" is a macros expansion of _Bool, which is a data-type. Personally, I find it annoying and odd-looking for "bool" to not be highlighted. True, it is not in the core "C language", but it is in the Standard C Library (stdbool.h).