Simplify encoding menu creation loop
Simplify and optimize creation of the sorted menus by taking advantage
of the fact they might already be partially sorted, and that we can at
least add one entry to each group in each run.

This goes from 4032 runs to 882, which while definitely not optimal for
adding 126 items, is probably totally good enough and don't warrant
duplicating the encoding array and sort it.

Though, such optimization doesn't matter as it's not what takes time in
this function, which is probably rather the widgets creation.