Remove anonymous structure identifier part in the symbol list
The counter used as a suffix of anonymous structures is shared by all
types such as structs, unions, enums, etc. This means that if there's
a single struct, union or enum in a single file, we get something like

anon_struct_1
anon_union_2
anon_enum_3

in the symbol list which is illogical from the user's perspective. Moreover,
if e.g. the union is part of a typedef, the typedef name is displayed
instead of anon_union_2 and we get

anon_struct_1
anon_enum_3

so it looks to the user that some numbers get skipped which looks strange.

Discard the anon_ prefix and the _number suffix and show just the type name.
Do a similar thing for Fortran types too.