Enable buffered draw on newer Gtk versions on OS X
Buffered draw seems to be faster and with newer Gtk versions
works correctly on HiDPI screens.

I haven't checked precisely since which Gtk version this works
(didn't work in 3.10, might have been fixed in a later version)
but it isn't as important because even with 3.16 the build
is highly unstable because of Gtk issues (Gtk 3.18 seems to
be promising though).