Add an API-like ctags layer
At the moment it is a bit hard to distinguish which of the functions
we are using belong to ctags. To make more explicit what we need
from ctags, wrap all ctags-related code and access ctags only using
this layer.

The interface from tm_ctags_wrappers.h can serve as a base for
proper ctags interface if ctags becomes a library.