Fix typename highlighting after using replace
At the moment we only re-colorize typenames at the visible part of the
screen (after full colorization when the file gets loaded). This works well
for normal editing but is insufficient for the replace operation as the
replaced text can be outside the visible area.

To fix this, set full_colourise=TRUE when replacing text, however, without
forcing the scintilla widget update so the coulourise type is only
remembered for the next rehighlight which happens after the file gets
parsed so it includes the change performed.