Use a format string instead of repeating almost the same string 9 times
This simplifies translations by being less redundant.

Note: label variable is not freed because `keybindings_set_item()`
currently assumes static strings but for plugins, so we need not free
the label right away.  It's not a real problem though, as this string
is required during the whole process' lifetime, so it'd use just as
much memory anyway.

Closes #953.