win32: do not -pthread in cross build script (for travis)
mingw-w64 in Ubuntu 12.04 doesn't like that option but it's not
necessary to get the cross build script working for travis. I didn't
verify that the resulting program actually works but that doesn't
matter for travis. We don't use the script for official win32 binaries
anyway.

The -pthread option is added to CFLAGS through pkg-config's .pc files.