api: correctly document ownership transfer
plugin_set_key_group() and _full() effectively take the ownership,
because the section_name is stored inside Geany.

Plugins usually pass a static string so it's not a problem. But GI-based
plugins might pass dynamic strings, so this behavior needs to be documented.

Fixes keybindings.conf being written with garbage (for a GI-based plugin).