Don't require plugin key group name and label strings to be static
Take a copy of the strings not to require them to live live as long as
the plugin does.

This is mostly useful for plugins implemented in dynamic languages
(e.g. through a plugin proxy), as most C plugins will use a static
string here; but it makes the API more straightforward and avoids odd
issues if any plugin doesn't use static strings here, even C ones.

Closes #1125.