Add keywords to the autocompletion list
Construct a dummy TMTag array containing keywords, filter them similarly
to workspace and global tags and show them together with normal tags
in the autocomplete popup.

Keyword groups that should be used for autocompletion are marked by
the "autocomplete" flag of HLKeyword and language mappings are updated
accordingly.

On the way fix config file of Cobol to contain valid keyword group.