Remove COMMITTERS
Geany is using git and github now so SF username isn't useful any more.
The current list of developers with push access differs from this list
too. The rest of the info is mentioned in AUTHORS and Frank's role as
a translation maintainer is mentioned in README.I18N (this info could
possibly be added to AUTHORS too).

This file doesn't contain any useful information any more and can be
removed.