Set document indent type when replacing tabs<->spaces in the whole document
When replacing tabs with spaces (or vice versa) in the whole document,
the user most probably wants to set also the indent type to the new
indent type. Do so.

In addition we have to take care of undo action which should undo both
the tabs in the document and the document indent type change - this can
be handled in a way similar to undoing EOL type change.

When replacing tabs/spaces only in a selection, don't change document
indent type in any way - in this case it's hard to say exactly what the
user wants.