Make UNDO_EOL/UNDO_INDENT always precede the attached Scintilla action
The undo/redo implementation of EOL/INDENT is asymmetrical now - for
undo we have UNDO_EOL/UNDO_INDENT at the top of the undo stack while
for redo we have UNDO_SCINTILLA at the top of the stack. Unify this by
adding UNDO_EOL/UNDO_INDENT inside the undo action after performing the
Scintilla change - we can do the same in redo and end up with similar
implementations of both undo and redo.