Port remaining toolbar settings to Stash
The UI for the icon size and style are a little tricky, because they
use radio buttons but spread them across two settings: the "system
default", and the specific values.  To handle this, we give a dummy
default to the "system default" in the specific setting, and handle
the separate "system default" setting *after* for it to override the
radio button choice.

This is somewhat tricky, but not worse than what was previously done,
where the specific setting wasn't loaded if the "system default" was
used, yet it was written (and actually used in the fallback path in
case the "system default" is missing), leading to using a potentially
random value.