Give '.copy-xx' filenames to cloned documents
The filename is generated by adding a suffix to it in the form of
'.copy-xx' where xx is a 2-digit number starting from 0 to 99.

Any generated filename in which the file it targets already exists, or
any generated filename that is already being used by another opened
document, gets skipped, and the iteration moves to the next number.  If
the iteration reaches 99 and no appropriate filename was still
generated, the cloned document is simply given with no filename or NULL.

No file or actual data is saved to disk yet.