tagmanager: new query module
The new module provides APIs to query tagmanager's loaded tags. The API is
aimed to be very flexible. A query object is created and filters can be added
to it which narrow down the resulting tags. The results can be retrieved by
executing the query, with optional post-exec sort and deduplication stages.

The goal is to eventually change the existing, use-case specific tag search
functions to use this new interface.