FXIME: Attempt at using the current (implicit) scope for symbol scope search
Try to resolve symbols for the current scope (if available) when
performing scope completion or searching for calltips.  This is based
on the assumption that for a symbol to be available in a scope, it has
to either be in a current or parent scope, or be part of an explicit
member.

This implementation is naive and done quickly, it's likely to have some
incorrect stuff in it.

Also, integration in scope completion is likely wrong.