Port "sidebar_pos" preference to use GSettings
This was already added to the schema initially to allow integration
to be done in the previous commit. Now it is actually used.

The setting is bound to the sidebar notebook page with the _SET, _GET,
and _GET_NO_CHANGES flags. This way the setting is kept in sync with
the active notebook page and the notebook page is initially set from
the setting but won't keep getting updated if the setting changes, for
example if another instance is using the same config file, changing
the sidebar page in one instance won't change it in the other.

For some reason not noted in the comments, changelog or commit message
the sidebar page was reset after plugins were loaded on startup in
b5884be2fa00f50880fb27a6fa5624499d57b4b7. This was removed since it is
now bound automatically. If that was previously done to save the tab
that plugins added, it wouldn't be right since not all plugins are
loaded at startup, nor in the same order, and plugins are unloaded
before the settings get saved anyway. If that was done to prevent
plugins hijacking the active sidebar page, it wouldn't be right since
it was only done at startup not when any plugin is loaded, and
additionally if a plugin simply appends a new tab page, it won't
activated anyway, unless they specifically make their new tab the
current page.