Port "tab_pos_*" preferences to GSettings
Due to a quirk in GSettings where enumerated types are stored as
GVariant strings rather than integers, and wanting to bind to the
"active" (integer) property of the combo boxes, they are bound with
mapping functions. These mapping functions can be re-used for any enum
with a GType as they are rather generic. There might be a better way
to do this.